refactor: split entities.py and players.py into focused modules
