+++ src/gem/extractors/players.py
@@ -9,10 +9,16 @@ Reference: examples/extraction_demo.py, refs/parser/src/main/java/opendota/Parse
 from __future__ import annotations
 
 import re
-from dataclasses import dataclass, field
 from typing import TYPE_CHECKING
 
 from gem.entities import Entity, EntityOp
+from gem.extractors._snapshots import (
+    _HERO_CLASS_PREFIX,
+    PlayerStateSnapshot,
+    PlayerTimeSeries,
+    _pos,
+    _snapshot_hero,
+)
 
 if TYPE_CHECKING:
     from gem.parser import ReplayParser
@@ -31,180 +37,10 @@ _NULL_HANDLE = 0xFFFFFF  # empty slot sentinel
 # Constants
 # ---------------------------------------------------------------------------
 
-_CELL_SIZE = 128  # Source 2 world units per grid cell
 _TEAM_RADIANT = 2
 _TEAM_DIRE = 3
-_HERO_CLASS_PREFIX = "CDOTA_Unit_Hero_"
 
-
-# ---------------------------------------------------------------------------
-# Helpers
-# ---------------------------------------------------------------------------
-
-
-def _pos(entity: Entity) -> tuple[float, float] | None:
-    """Return world (x, y) from cell+vec encoding on the entity.
-
-    Args:
-        entity: The entity to read coordinates from.
-
-    Returns:
-        ``(x, y)`` world coordinates, or ``None`` if any field is missing.
-    """
-    cell_x, ok_cx = entity.get_uint32("CBodyComponent.m_cellX")
-    cell_y, ok_cy = entity.get_uint32("CBodyComponent.m_cellY")
-    vec_x, ok_vx = entity.get_float32("CBodyComponent.m_vecX")
-    vec_y, ok_vy = entity.get_float32("CBodyComponent.m_vecY")
-    if not (ok_cx and ok_cy and ok_vx and ok_vy):
-        return None
-    return (cell_x * _CELL_SIZE + vec_x, cell_y * _CELL_SIZE + vec_y)
-
-
-def _snapshot_hero(entity: Entity, tick: int) -> PlayerStateSnapshot | None:
-    """Build a ``PlayerStateSnapshot`` from a hero entity.
-
-    Args:
-        entity: A ``CDOTA_Unit_Hero_*`` entity.
-        tick: Current game tick.
-
-    Returns:
-        A snapshot, or ``None`` if the player ID could not be resolved.
-    """
-    # m_nPlayerID (post-7.31) or m_iPlayerID (pre-7.31) — raw value is doubled;
-    # divide by 2 to get player slot 0-9. Reference: opendota/Parse.java getPlayerSlotFromEntity()
-    player_id, ok = entity.get_int32("m_nPlayerID")
-    if not ok:
-        player_id, ok = entity.get_int32("m_iPlayerID")
-    if not ok or player_id < 0:
-        return None
-    player_id //= 2
-
-    team, _ = entity.get_int32("m_iTeamNum")
-    level, _ = entity.get_int32("m_nCurrentLevel")
-    xp, _ = entity.get_int32("m_iCurrentXP")
-    hp, _ = entity.get_int32("m_iHealth")
-    max_hp, _ = entity.get_int32("m_iMaxHealth")
-    mana, _ = entity.get_float32("m_flMana")
-    max_mana, _ = entity.get_float32("m_flMaxMana")
-    lh, _ = entity.get_int32("m_iLastHitCount")
-    dn, _ = entity.get_int32("m_iDenies")
-
-    pos = _pos(entity)
-
-    # Convert entity class name to the canonical NPC name (camelCase → snake_case).
-    # "CDOTA_Unit_Hero_TemplarAssassin" → "npc_dota_hero_templar_assassin"
-    # "CDOTA_Unit_Hero_Nyx_Assassin"   → "npc_dota_hero_nyx_assassin" (already underscored)
-    # This matches dotaconstants keys and the combat log string table.
-    # Reference: refs/parser/Parse.java combatLogName2
-    _ending = entity.get_class_name()[len(_HERO_CLASS_PREFIX) :].replace("_", "")
-    _npc_name = "npc_dota_hero" + re.sub(r"([A-Z])", r"_\1", _ending).lower()
-    return PlayerStateSnapshot(
-        tick=tick,
-        player_id=player_id,
-        npc_name=_npc_name,
-        team=team,
-        level=level,
-        xp=xp,
-        gold=0,  # spendable gold — set by extractor from CDOTAPlayerController
-        total_earned_gold=0,  # cumulative — set by extractor from m_iTotalEarnedGold
-        net_worth=0,
-        lh=lh,
-        dn=dn,
-        hp=hp,
-        max_hp=max_hp,
-        mana=mana,
-        max_mana=max_mana,
-        x=pos[0] if pos else None,
-        y=pos[1] if pos else None,
-    )
-
-
-# ---------------------------------------------------------------------------
-# Data classes
-# ---------------------------------------------------------------------------
-
-
-@dataclass
-class PlayerStateSnapshot:
-    """A single per-player state sample taken at one tick.
-
-    Attributes:
-        tick: Game tick of this sample.
-        player_id: Player slot (0-9).
-        npc_name: Hero NPC name, e.g. ``"npc_dota_hero_axe"``.
-        team: Team number (2=Radiant, 3=Dire).
-        level: Hero level (1-30).
-        xp: Cumulative XP total.
-        gold: Spendable gold from ``CDOTAPlayerController``, or 0 if not read.
-        net_worth: Net worth from ``CDOTAPlayerController``, or 0 if not read.
-        total_earned_gold: Cumulative gold earned (``m_iTotalEarnedGold``), or 0 if not read.
-        total_earned_xp: Cumulative XP earned (``m_iTotalEarnedXP``), or 0 if not read.
-        lh: Last-hit count.
-        dn: Deny count.
-        hp: Current hit points.
-        max_hp: Maximum hit points.
-        mana: Current mana.
-        max_mana: Maximum mana.
-        x: World x coordinate, or ``None`` if unavailable.
-        y: World y coordinate, or ``None`` if unavailable.
-        ability_levels: Ability name → level mapping for learned abilities.
-    """
-
-    tick: int
-    player_id: int
-    npc_name: str
-    team: int
-    level: int
-    xp: int
-    gold: int
-    net_worth: int
-    lh: int
-    dn: int
-    hp: int
-    max_hp: int
-    mana: float
-    max_mana: float
-    x: float | None
-    y: float | None
-    total_earned_gold: int = 0
-    total_earned_xp: int = 0
-    ability_levels: dict[str, int] = field(default_factory=dict)
-
-
-@dataclass
-class PlayerTimeSeries:
-    """Time-series data for one player, aggregated from snapshots.
-
-    Attributes:
-        player_id: Player slot (0-9).
-        ticks: Tick values for each sample.
-        gold_t: Spendable gold at each sample tick.
-        total_earned_gold_t: Cumulative total earned gold at each sample tick.
-        total_earned_xp_t: Cumulative total earned XP at each sample tick.
-        net_worth_t: Net worth at each sample tick.
-        lh_t: Last-hit count at each sample tick.
-        dn_t: Deny count at each sample tick.
-        xp_t: Cumulative XP at each sample tick.
-        hp_t: Current hit points at each sample tick.
-        mana_t: Current mana at each sample tick.
-        x_t: World x coordinate at each sample tick (``None`` if unavailable).
-        y_t: World y coordinate at each sample tick (``None`` if unavailable).
-    """
-
-    player_id: int
-    ticks: list[int] = field(default_factory=list)
-    gold_t: list[int] = field(default_factory=list)
-    total_earned_gold_t: list[int] = field(default_factory=list)
-    total_earned_xp_t: list[int] = field(default_factory=list)
-    net_worth_t: list[int] = field(default_factory=list)
-    lh_t: list[int] = field(default_factory=list)
-    dn_t: list[int] = field(default_factory=list)
-    xp_t: list[int] = field(default_factory=list)
-    hp_t: list[int] = field(default_factory=list)
-    mana_t: list[float] = field(default_factory=list)
-    x_t: list[float | None] = field(default_factory=list)
-    y_t: list[float | None] = field(default_factory=list)
-
+__all__ = ["PlayerExtractor", "PlayerStateSnapshot", "PlayerTimeSeries"]
 
 # ---------------------------------------------------------------------------
 # Extractor
